requirements.txt.txt
flask
mysql-connector-python
python-dotenv
requests
